moved DataType into affordance package
