+++ src/main/java/de/escalon/hypermedia/affordance/DataType.java
@@ -8,7 +8,7 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-package de.escalon.hypermedia;
+package de.escalon.hypermedia.affordance;
 
 import org.omg.CORBA.Current;
 import org.springframework.format.number.NumberFormatter;
